Move AddLastCommitCache to a common nogogit and gogit file
